振動メソッドをHapticsファイルに移動
